Fixes grounded jumping
Ground-checking raycast is now created 0.1 meters above the foot of the vulture (the base vulture gameobject). The raycast extends downwards 0.2 meters and hits platforms that it would have ignored due to already colliding with them beforehand.

In addition, platform collision with tagged gameobjects is checked by the base vulture gameobject (which now contains both the vulture's rigidbody and the collider) to ensure that jumping is done when both a platform is in contact with the player and the raycast can detect something below the player.

Also, VultureStateMachine now disables State classes not in use. VultureStateClass also contains an event action that reassigns the state for the state machine. A ducking jump and jump power based on the amount of time the jump button has been held is up next for implementation.