Adds tilting platform, collision fix, and timer
Tilting platform prefab added, with script that reorients the platform once the player stops colliding with it. Vulture Object's platform collision edited to set platformContact to false at each fixedUpdate, with the state script added earlier to the execution order to obtain platformContact info before it's set to false. Carrion timer also added to UI, with the canvas added to the LevelManager Object and timer countdown added to CarrionSpawner scripts to begin the carrion countdown and destroy carrion/reset the timer, depending on if the player collected the carrion in time.